Refactor the architecture
