+++ src/config/grid_search_config.py
@@ -1,18 +1,9 @@
 import copy
 from pathlib import Path
 
-from git import Repo
-from omegaconf import OmegaConf
 from sklearn.model_selection import ParameterGrid
 
-
-def create_config(config_name: str):
-    config = load_yaml_file_content(Path("config") / config_name)
-
-    if config["is_logging_commit_info"]:
-        config = update_config_with_commit_name_and_hash(config)
-
-    return config
+from src.config.load import load_yaml_file_content
 
 
 def create_config_combinations_sorted_by_dataset(config: dict) -> list[dict]:
@@ -57,35 +48,3 @@ def overrode_config_with_grid_search_config(config: dict, grid_search_config: di
 def create_parameter_grid(config: dict, grid_search_config: dict) -> ParameterGrid:
     config = {key: [value] if key not in grid_search_config else value for key, value in config.items()}
     return ParameterGrid([config])
-
-
-def load_yaml_file_content(file_path: Path) -> dict:
-    omega_config = OmegaConf.load(file_path)
-    return OmegaConf.to_container(omega_config, resolve=True)
-
-
-def update_config_with_commit_name_and_hash(config: dict) -> dict:
-    repo = Repo()
-
-    if repo.is_dirty(untracked_files=True):
-        unstaged_files_paths = repo.untracked_files + [diff.a_path for diff in repo.index.diff(None)]
-        staged_files_paths = [diff.a_path for diff in repo.index.diff(repo.head.commit)]
-
-        error_message = ("Experiments might not be reproducible if your working tree isn't clean."
-                         f"Please commit your changes in {unstaged_files_paths + staged_files_paths}")
-        raise Exception(error_message)
-
-    config["commit_hash"] = get_commit_hash(repo)
-    config["commit_name"] = get_commit_name(repo)
-
-    return config
-
-
-def get_commit_hash(repo: Repo) -> str:
-    latest_commit = repo.head.commit
-    return latest_commit.hexsha
-
-
-def get_commit_name(repo: Repo) -> str:
-    latest_commit = repo.head.commit
-    return latest_commit.message
