+++ src/data/create_datasets.py
@@ -1,8 +1,8 @@
-from dataset.blob import generate_blob_datasets
-from dataset.mnist import load_mnist
-from dataset.moon import generate_moon_datasets
-from dataset.moon_and_blob import generate_moon_and_blob_datasets
-from dataset.mtpl import load_MTPL
+from src.data.dataset.blob import generate_blob_datasets
+from src.data.dataset.mnist import load_mnist
+from src.data.dataset.moon import generate_moon_datasets
+from src.data.dataset.moon_and_blob import generate_moon_and_blob_datasets
+from src.data.dataset.mtpl import load_MTPL
 
 
 def create_datasets(config: dict):
