+++ src/data/dataset/mnist.py
