+++ src/result/prevent_running_completed_job.py
@@ -1,0 +1,66 @@
+def is_job_already_done(project_name, task_dict):
+    """
+    Verifies if a hyperparameter combination has already been tested.
+    Args:
+        project_name (str): The name of the wandb project;
+        task_dict (dictionary): the dictionary containing the current hyperparameters combination;
+    Return:
+        bool, whether the combination has already been tested or not
+    """
+    cnt_nw = 0
+    new, keys = [], []
+    for key in task_dict:
+        keys.append(key)
+    keys.sort()
+    for key in keys:
+        new.append(str(task_dict[key]))
+    try:
+        with open("results/" + str(project_name) + ".txt", "r") as tes:
+            tess = [line.strip().split("\t") for line in tes]
+        tes.close()
+        for i in range(len(tess)):
+            if tess[i][:len(new)] == new:
+                cnt_nw += 1
+    except FileNotFoundError:
+        file = open("results/" + str(project_name) + ".txt", "a")
+        for key in keys:
+            file.write(key + "\t")
+        file.write("train_acc" + "\t")
+        file.write("valid_acc" + "\t")
+        file.write("test_acc" + "\t")
+        file.write("bound_lin" + "\t")
+        file.write("bound_hyp" + "\t")
+        file.write("bound_kl" + "\t")
+        file.write("bound_mrch" + "\t")
+        file.write("n_sigma" + "\t")
+        file.write("n_Z" + "\n")
+        file.close()
+    return cnt_nw > 0
+
+
+def write(file_name, task_dict, hist, best_epoch):
+    """
+    Writes in a .txt file the hyperparameters and results of a training of the BGN algorithm
+        on a given dataset.
+    Args:
+        file_name (str): The name of the .txt file to write into;
+        task_dict (dictionary): the dictionary containing the current hyperparameters combination;
+        hist (dictionary): A dictionary that keep track of training metrics.
+        best_epoch (int): best epoch.
+    """
+    keys = []
+    for key in task_dict:
+        keys.append(key)
+    keys.sort()
+    file = open("results/" + str(file_name) + ".txt", "a")
+    for key in keys:
+        file.write(str(task_dict[key]) + "\t")
+    file.write(str(hist["train_acc"][best_epoch].item()) + "\t")
+    file.write(str(hist["valid_acc"][best_epoch].item()) + "\t")
+    file.write(str(hist["test_acc"][best_epoch].item()) + "\t")
+    file.write(str(hist["bound_lin"][best_epoch].item()) + "\t")
+    file.write(str(hist["bound_hyp"][best_epoch].item()) + "\t")
+    file.write(str(hist["bound_kl"][best_epoch]) + "\t")
+    file.write(str(hist["bound_mrch"][best_epoch]) + "\t")
+    file.write("\n")
+    file.close()
