+++ src/utils.py
@@ -1,9 +1,24 @@
+import random
+
+import numpy as np
+import torch
+from matplotlib import pyplot as plt
+
+plt.switch_backend("agg")
+
+
+def set_random_seed(seed: int) -> None:
+    np.random.seed(seed)
+    torch.manual_seed(seed)
+    random.seed(seed)
+
+
 def create_run_name(task_dict: dict) -> str:
     run_name_content = task_dict["run_name_content"]
 
     if not run_name_content:
         return "Default name"
-    
+
     run_name_elements = []
     for parameter_name in run_name_content:
         run_name_elements.append(f"{parameter_name.replace('_', '-')}={task_dict[parameter_name]}")
